// frontend/src/app/enums/select-mapping.ts
export enum RoleEnum {
  MENTORADO = 'mentorado',
  MENTOR = 'mentor',
}

export enum DepartmentEnum {
  TECNOLOGIA = 'Tecnologia',
  PRODUTO = 'Produto',
  MARKETING = 'Marketing',
  PESSOAS_CULTURA = 'Pessoas e Cultura',
}

export enum AreaEnum {
  DEV_BACK_END = 'DevBack-end',
  QA = 'QA',
  PRODUTO = 'Produto',
  UX_UI_DESIGNER = 'UX/UIDesigner',
  DATA_ENGINEER = 'Data Engineer',
  DEV_FRONT_END = 'DevFront-end',
  DEV_MOBILE = 'DevMobile',
}

export enum FormationEnum {
  SIM = 'Sim',
  NAO = 'Não',
}

export enum TaskPriorityEnum {
  BAIXA = 'baixa',
  MEDIA = 'media',
  ALTA = 'alta',
}

export enum TaskStatusEnum {
  PENDENTE = 'pendente',
  ANDAMENTO = 'andamento',
  CONCLUIDA = 'concluida',
}